divide GA into two file
