+++ tuner/GA.py
@@ -1,15 +1,16 @@
-import os
-import math
-import sys
-import logging
 import argparse
+import datetime
+import logging
+import math
+import os
+import sys
 from tqdm import tqdm
 
-import torch
-import datetime
 import numpy as np
 import pandas as pd
+import torch
 
+from score_function import *
 from knobs import generate_config
 import utils
 
@@ -35,49 +36,6 @@ args = parser.parse_args()
 
 if not os.path.exists('save_knobs'):
     assert "Do this file after running main.py"
-
-def mse_loss(target, predict):
-    return np.array([(target[:,i]-predict[:,i])**2 for i in range(2)]).sum(axis=0)
-
-def ATR_loss(default, predict, weight):
-    return sum([((-1**i)*weight[i]*(predict[:,i]-default[:,i]))/default[:,i] for i in range(len(weight))])
-
-def throughput_loss(default, predict):
-    loss = []
-    for d, p in zip(default, predict):
-        if p-d>0:
-            loss.append(math.log10(p-d) + 1)
-        elif p-d == 0:
-            loss.append(0)
-        elif p-d<0:
-            loss.append(-(math.log10(abs(p-d)) + 1))
-    return np.array(loss)
-
-def latency_loss(default, predict):
-    loss = []
-    for d, p in zip(default, predict):
-        if d-p>0:
-            loss.append(-pow(2,d-p))
-        elif d-p==0:
-            loss.append(0)
-        elif d-p<0:
-            loss.append(pow(2,d-p))
-    return np.array(loss)
-
-def DRT_loss(default, predict, weight):
-    losses = [throughput_loss, latency_loss]
-    return sum([weight[i]*losses[i](default[:,i], predict[:,i]) for i in range(len(weight))])
-
-def throughput_new_loss(default, predict):
-    return (predict/default-1)*100
-
-def latency_new_loss(default, predict):
-    return (1-(predict/default))
-
-def DRT_new_loss(default, predict, weight):
-    losses = [throughput_new_loss, latency_new_loss]
-    return sum([weight[i]*losses[i](default[:,i], predict[:,i])*((-1)**i) for i in range(len(weight))])
-
 
 print("======================MAKE GA LOGGER====================")
 logger, log_dir = utils.get_logger(os.path.join('./GA_logs'))
@@ -108,9 +66,8 @@ def main():
     if args.sk==' ':
         args.sk = args.path
 
+    TUNE_EM = ["Totals_Ops/sec","Totals_p99_Latency"]
     top_k_knobs = np.load(os.path.join('./save_knobs',args.sk,f"knobs_{args.topk}.npy"))
-    #import utils
-    #predict_save_path = utils.make_date_dir('./save_predicts')
 
     if args.model_mode == 'single':
         model = RedisSingleDNN(args.topk+5,2)
@@ -125,13 +82,18 @@ def main():
         models[0].load_state_dict(torch.load(os.path.join('./model_save',args.path,'Totals_Ops_sec_{}.pt'.format(args.num[0]))))
         models[1].load_state_dict(torch.load(os.path.join('./model_save',args.path,'Totals_p99_Latency_{}.pt'.format(args.num[1]))))
         fitness_function = double_fitness_function
+    # elif args.model_mode == 'double':
+    #     models = [RedisSingleDNN(args.topk,1), RedisSingleDNN(args.topk,1)]
+    #     models[0].load_state_dict(torch.load(os.path.join('./model_save',args.path,'Totals_Ops_sec_{}.pt'.format(args.num[0]))))
+    #     models[1].load_state_dict(torch.load(os.path.join('./model_save',args.path,'Totals_p99_Latency_{}.pt'.format(args.num[1]))))
+    #     fitness_function = double_fitness_function
 
     if args.model_mode == 'single' or args.model_mode == 'twice':
-        pruned_configs, external_data, default, scaler_X, scaler_y = prepareForGA(args,top_k_knobs)
-        temp_configs = pd.concat([pruned_configs, external_data], axis=1)
-        temp_configs = temp_configs.sort_values(["Totals_Ops/sec","Totals_p99_Latency"], ascending=[False,True])
-        target = temp_configs[["Totals_Ops/sec","Totals_p99_Latency"]].values[0]
-        configs = temp_configs.drop(columns=["Totals_Ops/sec","Totals_p99_Latency"])
+        pruned_configs, external_data, default, scaler_X, scaler_y = prepareForGA(args, top_k_knobs)
+        temp_configs = pd.concat([pruned_configs, external_data], axis = 1)
+        temp_configs = temp_configs.sort_values(TUNE_EM, ascending = [False, True])
+        target = temp_configs[TUNE_EM].values[0]
+        configs = temp_configs.drop(columns = TUNE_EM)
         current_solution_pool = configs[:args.n_pool].values
         target = np.repeat([default], args.n_pool, axis = 0)
 
@@ -141,45 +103,62 @@ def main():
             default: [ops_deafult, lat_deafult]
             scaler_y: [scaler_ops, scaler_lat]
         """
-        pruned_configs, external_datas, defaults, scaler_X, scaler_ys = double_prepareForGA(args,top_k_knobs)
-        temp_configs = [pd.concat([pruned_configs, external_data], axis=1) for external_data in external_datas]
-        temp_configs[0] = temp_configs[0].sort_values(["Totals_Ops/sec"], ascending=[False])
-        temp_configs[1] = temp_configs[1].sort_values(["Totals_p99_Latency"], ascending=[True])
-        targets = [None, None]
-        targets[0] = temp_configs[0][["Totals_Ops/sec"]].values[0]
-        targets[1] = temp_configs[1][["Totals_p99_Latency"]].values[0]
-        configs = [None, None]
-        configs[0] = temp_configs[0].drop(columns=["Totals_Ops/sec"])
-        configs[1] = temp_configs[1].drop(columns=["Totals_p99_Latency"])
+        pruned_configs, external_datas, defaults, scaler_X, scaler_ys = double_prepareForGA(args, top_k_knobs)
+        temp_configs = [pd.concat([pruned_configs, external_data], axis = 1) for external_data in external_datas]
+        targets = []
+        configs = []
+        for i in range(len(TUNE_EM)):
+            temp_configs[i] = temp_configs[i].sort_values(TUNE_EM[i], ascending = [i])
+
+        for i in range(len(TUNE_EM)):
+            targets.append(temp_configs[i][[TUNE_EM[i]]].values[0])
+
+        for i in range(len(TUNE_EM)):
+            configs.append(temp_configs[i].drop(columns = [TUNE_EM[i]]))
+
         current_solution_pools = [config[:args.n_pool].values for config in configs]
         targets = [np.repeat([default], args.n_pool, axis = 0) for default in defaults]
         
     #losses = [throughput_loss, latency_loss]
     losses = [throughput_new_loss, latency_new_loss]
     n_configs = top_k_knobs.shape[0]
+    #set remain ratio
     n_pool_half = args.n_pool//2
+    #mutation ratio
     mutation = int(n_configs*0.5)
+
     ops_predicts = []
     lat_predicts = []
     for i in tqdm(range(args.n_generation)):
         if args.model_mode == 'single' or args.model_mode == 'twice':
+            #fitness function
             scaled_pool = scaler_X.transform(current_solution_pool)
             predicts = fitness_function(scaled_pool, args, model)
             fitness = scaler_y.inverse_transform(predicts)
+
+            #save preidct ops and latency
             ops_predicts.append(np.max(fitness[:,0]))
             lat_predicts.append(np.min(fitness[:,1]))
+
             #idx_fitness = ATR_loss(target, fitness,[0.5,0.5])
+            #score function and sort by score
             idx_fitness = DRT_new_loss(target, fitness,[0.5,0.5])
             sorted_idx_fitness = np.argsort(idx_fitness)[n_pool_half:]
-            best_solution_pool = current_solution_pool[sorted_idx_fitness,:]
+            best_solution_pool = current_solution_pool[sorted_idx_fitness, :]
+
             if i % 1000 == 999:
                 logger.info(f"[{i+1:3d}/{args.n_generation:3d}] best fitness: {max(idx_fitness)}")
-            pivot = np.random.choice(np.arange(1,n_configs))
+
+            #random select crossover ratio
+            pivot = np.random.choice(np.arange(1, n_configs))
             new_solution_pool = np.zeros_like(best_solution_pool)
             for j in range(n_pool_half):
+                #crossover
                 new_solution_pool[j][:pivot] = best_solution_pool[j][:pivot]
                 new_solution_pool[j][pivot:n_configs] = best_solution_pool[n_pool_half-1-j][pivot:n_configs]
                 new_solution_pool[j][n_configs:] = current_solution_pool[0][n_configs:]
+
+                #mutation
                 import utils, random
                 random_knobs = utils.make_random_option(top_k_knobs)
                 knobs = list(random_knobs.values())
@@ -188,30 +167,43 @@ def main():
                 random_knob_index = random_knob_index[:mutation]
                 for k in range(len(random_knob_index)):
                     new_solution_pool[j][random_knob_index[k]] = knobs[random_knob_index[k]]
+
+            #new solution pool for next generation
             current_solution_pool = np.vstack([best_solution_pool, new_solution_pool])
+
         elif args.model_mode == 'double':
+            #fitness function
             index = i%2
             scaled_pool = scaler_X.transform(current_solution_pools[index])
             predicts = fitness_function(scaled_pool, args, models[index])
             fitness = scaler_ys[index].inverse_transform(predicts)
+
+            #save preidct ops and latency
             if index:
                 lat_predicts.append(np.min(fitness))
             else:
                 ops_predicts.append(np.max(fitness))
+            
+            #score function and sort by score
             idx_fitness = np.squeeze(losses[index](targets[index], fitness))
-            #idx_fitness = ATR_loss(targets[index], fitness,[0.5])
             sorted_idx_fitness = np.argsort(idx_fitness)[n_pool_half:]
             best_solution_pool = current_solution_pools[index][sorted_idx_fitness,:]
+
             if i % 1000 == 998:
                 logger.info(f"[{i+1:3d}/{args.n_generation:3d}] best fitness: {max(idx_fitness)}")
             if i % 1000 == 999:
                 logger.info(f"[{i+1:3d}/{args.n_generation:3d}] best fitness: {max(idx_fitness)}")
+            
+            #random select crossover ratio
             pivot = np.random.choice(np.arange(1,n_configs))
             new_solution_pool = np.zeros_like(best_solution_pool)
             for j in range(n_pool_half):
+                #crossover
                 new_solution_pool[j][:pivot] = best_solution_pool[j][:pivot]
                 new_solution_pool[j][pivot:n_configs] = best_solution_pool[n_pool_half-1-j][pivot:n_configs]
                 new_solution_pool[j][n_configs:] = current_solution_pools[index][0][n_configs:]
+                
+                #mutation
                 import utils, random
                 random_knobs = utils.make_random_option(top_k_knobs)
                 knobs = list(random_knobs.values())
@@ -220,19 +212,26 @@ def main():
                 random_knob_index = random_knob_index[:mutation]
                 for k in range(len(random_knob_index)):
                     new_solution_pool[j][random_knob_index[k]] = knobs[random_knob_index[k]]
+
+            #new solution pool for next generation        
             current_solution_pools[index] = np.vstack([best_solution_pool, new_solution_pool])
     np.save(os.path.join('save_predicts',f'{args.persistence}_{args.n_pool}_{args.target}_ops.npy'),np.array(ops_predicts))
     np.save(os.path.join('save_predicts',f'{args.persistence}_{args.n_pool}_{args.target}_lat.npy'),np.array(lat_predicts))
     
+    #generation_config by GA results
     final_solution_pool = pd.DataFrame(best_solution_pool)
     logger.info(top_k_knobs)
     logger.info(final_solution_pool)
     top_k_config_path, name, connect = generate_config(args, top_k_knobs, final_solution_pool)
+
+    #connect other server to run memtier_benchmarks
     if connect:
         server_connection(args, top_k_config_path, name)
     else:
         logger.info("Because appednfsync is 'always', Fin GA")
         return 0
+
+    #save results
     i = 0
     today = datetime.datetime.now()
     name = 'result_'+args.persistence+'-'+today.strftime('%Y%m%d')+'-'+'%02d'%i+'.csv'
